# services/auth/openapi/identity.yaml
...
openapi: 3.1.0
info:
  title: NEXUS Identity API
  version: 0.1.0
  description: |
    Authentication and credential lifecycle for NEXUS Commerce OS.

    Sessions belong to NEXUS. External identity providers authenticate a person;
    they never own an application session. Every authentication mechanism
    converges on the same Session aggregate, which owns creation, rotation,
    refresh-token lifecycle, reuse detection, revocation and device binding.
servers:
  - url: /v1
    description: Version prefix
...
paths:
  /auth/register:
    post:
      operationId: registerUser
      tags: [registration]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: The account was created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegisteredUser' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /auth/login:
    post:
      operationId: login
      tags: [session]
      summary: Exchange a password for a NEXUS session
      description: |
        On success a session is started on the NEXUS Session aggregate. A wrong
        password and an unknown account are deliberately indistinguishable.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: A session was started.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionIssued' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /auth/session/refresh:
    post:
      operationId: refreshSession
      tags: [session]
      summary: Rotate a refresh token
      description: |
        The presented token is consumed and a replacement issued. Presenting an
        already-rotated token is treated as theft: the session is revoked.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: The session was refreshed and the token rotated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionIssued' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /auth/email/verify:
    post:
      operationId: verifyEmail
      tags: [email]
      summary: Consume an email-verification link
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TokenRequest' }
      responses:
        '204':
          description: The address is verified.
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /auth/password/reset-requests:
    post:
      operationId: requestPasswordReset
      tags: [password]
      summary: Ask for a password-reset link
      description: |
        Always answers 202, whether or not the address exists and whether or not
        the message could be delivered. Any other behaviour would let a caller
        enumerate accounts.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordResetRequest' }
      responses:
        '202':
          description: If the address matches an account, a link has been sent.
        '400': { $ref: '#/components/responses/Problem' }
  /auth/password/reset:
    post:
      operationId: resetPassword
      tags: [password]
      summary: Consume a reset link and set a new password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordResetConfirm' }
      responses:
        '204':
          description: The password was changed and every session revoked.
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /auth/logout:
    post:
      operationId: logout
      tags: [session]
      summary: End the current session
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmptyRequest' }
      responses:
        '204':
          description: The session is revoked.
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /auth/logout-all:
    post:
      operationId: logoutAll
      tags: [session]
      summary: End every session for the user
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LogoutAllRequest' }
      responses:
        '204':
          description: Every matching session is revoked.
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
  /auth/password/change:
    post:
      operationId: changePassword
      tags: [password]
      summary: Change the password of the signed-in account
      description: Every session is revoked afterwards, including this one.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '204':
          description: The password was changed.
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /auth/email/verification-requests:
    post:
      operationId: requestEmailVerification
      tags: [email]
      summary: Send a fresh verification link to the signed-in account
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmptyRequest' }
      responses:
        '202':
          description: A verification link has been sent.
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '502': { $ref: '#/components/responses/Problem' }
  /auth/passkeys/registration/start:
    post:
      operationId: startPasskeyRegistration
      tags: [passkey]
      summary: Begin registering a passkey
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmptyRequest' }
      responses:
        '200':
          description: A single-use challenge for the browser ceremony.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PasskeyChallenge' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /auth/passkeys/registration/complete:
    post:
      operationId: completePasskeyRegistration
      tags: [passkey]
      summary: Finish registering a passkey
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompletePasskeyRegistrationRequest' }
      responses:
        '201':
          description: The credential is registered.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegisteredPasskey' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /auth/passkeys/authentication/start:
    post:
      operationId: startPasskeyAuthentication
      tags: [passkey]
      summary: Begin signing in with a passkey
      description: Omit the address entirely for a discoverable-credential sign-in.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartPasskeyAuthenticationRequest' }
      responses:
        '200':
          description: A single-use challenge for the browser ceremony.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PasskeyChallenge' }
        '400': { $ref: '#/components/responses/Problem' }
  /auth/passkeys/authentication/complete:
    post:
      operationId: completePasskeyAuthentication
      tags: [passkey, session]
      summary: Finish signing in with a passkey
      description: |
        Converges on the same NEXUS Session lifecycle as password login and
        returns the identical credential model. The WebAuthn assertion is never
        itself a request credential.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompletePasskeyAuthenticationRequest' }
      responses:
        '200':
          description: A session was started.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionIssued' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /auth/passkeys:
    get:
      operationId: listPasskeys
      tags: [passkey]
      summary: List the signed-in user's active passkeys
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: The active credentials.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PasskeyList' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /auth/passkeys/revoke:
    post:
      operationId: revokePasskey
      tags: [passkey]
      summary: Revoke one of the signed-in user's passkeys
      description: Refused when it is the last usable authentication factor.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RevokePasskeyRequest' }
      responses:
        '204':
          description: The credential is revoked.
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /auth/oidc/start:
    post:
      operationId: startOidcLogin
      tags: [federation]
      summary: Begin a federated sign-in
      description: |
        Unauthenticated for a sign-in. Supplying a bearer token instead begins a
        link flow, attaching the provider to the signed-in account.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartOidcRequest' }
      responses:
        '200':
          description: State, nonce and PKCE verifier for the authorization request.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OidcStarted' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /auth/oidc/complete:
    post:
      operationId: completeOidcLogin
      tags: [federation, session]
      summary: Finish a federated sign-in
      description: |
        State and nonce are mandatory and validated, as are the provider issuer
        and audience. The provider tokens never become NEXUS request
        credentials: a NEXUS Session is created and its own credentials returned.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteOidcRequest' }
      responses:
        '200':
          description: A session was started.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OidcSessionIssued' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /auth/federated-identities:
    get:
      operationId: listFederatedIdentities
      tags: [federation]
      summary: List the signed-in user's linked providers
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: The active links.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FederatedIdentityList' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /auth/federated-identities/unlink:
    post:
      operationId: unlinkFederatedIdentity
      tags: [federation]
      summary: Unlink a provider from the signed-in account
      description: Refused when it is the last usable authentication factor.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UnlinkFederatedIdentityRequest' }
      responses:
        '204':
          description: The provider is unlinked.
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The short-lived access token from `login` or `session/refresh`. The
        refresh token is never accepted here.
  responses:
    Problem:
      description: RFC 9457 problem detail.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password, displayName]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 12, maxLength: 1024 }
        displayName: { type: string, minLength: 1, maxLength: 120 }
        locale:
          type: string
          description: BCP-47 tag. Defaults to en-US.
          pattern: '^[A-Za-z]{2,3}(-[A-Za-z0-9]{2,8})*$'
        residencyRegion:
          type: string
          description: Data-residency region (ADR-0007). Defaults to us-east-1.
          pattern: '^[a-z]{2}-[a-z]+-[0-9]$'
    RegisteredUser:
      type: object
      additionalProperties: false
      required: [id, email, emailVerified]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        emailVerified: { type: boolean }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, maxLength: 254 }
        password: { type: string, maxLength: 1024 }
        deviceBinding:
          type: string
          description: Opaque device reference to bind the session to.
          maxLength: 200
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken: { type: string, minLength: 1, maxLength: 512 }
    TokenRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token: { type: string, minLength: 1, maxLength: 512 }
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: { type: string, maxLength: 254 }
    PasswordResetConfirm:
      type: object
      additionalProperties: false
      required: [token, newPassword]
      properties:
        token: { type: string, minLength: 1, maxLength: 512 }
        newPassword: { type: string, minLength: 12, maxLength: 1024 }
    SessionIssued:
      type: object
      additionalProperties: false
      required: [session, accessToken, expiresAt, refreshToken]
      properties:
        session: { $ref: '#/components/schemas/Session' }
        accessToken:
          type: string
          description: |
            Short-lived JWT presented on every authenticated request. Derived
            from the session; the session stays authoritative for revocation.
        expiresAt:
          type: string
          format: date-time
          description: When the access token lapses.
        refreshToken:
          type: string
          description: |
            Rotating secret, returned once, used only to obtain a new access
            token. Never presented as a request credential.
...
    EmptyRequest:
      type: object
      additionalProperties: false
      description: No fields; the principal comes from the bearer token.
    LogoutAllRequest:
      type: object
      additionalProperties: false
      properties:
        exceptCurrent:
          type: boolean
          description: Keep this session alive — "sign out everywhere else".
...